# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SahamRadar AI - Indonesian Stock Screener API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Get user watchlist
      parameters:
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add stock to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /watchlist/{ticker}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove stock from watchlist
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
components:
  schemas:
...
    Stock:
      type: object
      required: [ticker, name, sector, currentPrice, priceChange, priceChangePct, volume, totalScore, label]
      properties:
        id:
          type: integer
        ticker:
          type: string
        name:
          type: string
        sector:
          type: string
        currentPrice:
          type: number
        priceChange:
          type: number
        priceChangePct:
          type: number
        volume:
          type: integer
        avgVolume:
          type: ["integer", "null"]
        marketCap:
          type: ["number", "null"]
        trendScore:
          type: number
        momentumScore:
          type: number
        volumeScore:
          type: number
        liquidityScore:
          type: number
        fundamentalScore:
          type: number
        valuationScore:
          type: number
        riskScore:
          type: number
        totalScore:
          type: number
        label:
          type: string
        updatedAt:
          type: string
...
    WatchlistItem:
      type: object
      required: [id, userId, ticker, addedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        ticker:
          type: string
        addedAt:
          type: string
        stock:
          $ref: "#/components/schemas/Stock"
    WatchlistInput:
      type: object
      required: [userId, ticker]
      properties:
        userId:
          type: string
        ticker:
          type: string
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...